pet file

+++ src/main/java/com/care/am/service/medi/mediServiceImpl.java
@@ -164,7 +164,10 @@ public class mediServiceImpl implements mediService{
 		
 		//file 
 		String dbImg = mm.getMedi(dto.getmId()).getmPhoto(); //디비에 저장되어있는 사진 이름
-		mfs.deleteImage(dbImg);
+		if(!dbImg.equals("mediDefault.jpg")) {
+			mfs.deleteImage(dbImg);
+		}
+		
 		String originName = file.getOriginalFilename();	
 		
 		if(originName=="") { //파일 선택이 없다면  //
@@ -179,12 +182,14 @@ public class mediServiceImpl implements mediService{
 		String msg ="", url="";
 		if(result==1) {
 			mfs.deleteImage(originName);
+			System.out.println("=-=-=-=-=-==-삭제할 파일이름: "+originName);
 			return GetMessage.getMessage("정보가 수정되었습니다!", "/am/mediInfo?id=" + dto.getmId());
 		}else {
 			mfs.deleteImage(dto.getmPhoto());
 			return GetMessage.getMessage( "정보 수정에 실패했습니다.", "/am/mediModify?id=" + dto.getmId());
 		}
 	}
+	
 	public String mediDelete(mediDTO dto, String pw) {
 		dto = mm.getMedi(dto.getmId());
 		 int result =0 ;
